@Override added, HandlerManager moved to CompositeEventHandler
