--- src/main/java/ru/sbt/mipt/smarthome/HandlerManager.java
@@ -1,7 +1,0 @@
-package ru.sbt.mipt.smarthome;
-
-import ru.sbt.mipt.smarthome.events.SensorEvent;
-
-public interface HandlerManager {
-    boolean processEvent(SensorEvent event);
-}
