+++ src/main/java/ru/sbt/mipt/smarthome/events/LightOff.java
@@ -10,6 +10,7 @@ public class LightOff implements SensorEvent {
         this.deviceId = deviceId;
     }
 
+    @Override
     public String getComponentId() {
         return deviceId;
     }
